// src/frontend/lib/types.ts
export interface SearchResult {
  id: string;
  title: string;
  poster: string;
  episodes: string[];
  episodes_titles: string[];
  source: string;
  source_name: string;
  class?: string;
  year: string;
  desc?: string;
  type_name?: string;
  douban_id?: number;
  remarks?: string; // 备注信息（如"已完结"、"更新至20集"等）
}